edit : timeCard를 헤드리스 ui로 수정 적용
