+++ src/features/plan/components/TimeCard/TimeCardView.tsx
@@ -1,0 +1,19 @@
+import { TimeCardType } from '@/types/cardTypes'
+import { Clock } from 'lucide-react'
+
+// 재사용 가능한 Card 컴포넌트 (시간 선택용)
+export const TimeCard = ({ time, isActive, onClick }: TimeCardType) => {
+  return (
+    <button
+      onClick={onClick}
+      className={`flex-1 p-4 rounded-xl transition-all duration-300 ${
+        isActive
+          ? 'bg-indigo-600 text-white shadow-md'
+          : 'bg-white text-gray-700 hover:shadow-sm border-2 border-gray-100'
+      }`}
+    >
+      <Clock className="w-5 h-5 mx-auto mb-1" />
+      <p className="text-sm font-semibold">{time}분</p>
+    </button>
+  )
+}
